flutter fix offers logic. Add group users by id

+++ lib/data/api/models/user_rating_model.dart
@@ -7,7 +7,7 @@ class UserRatingModel {
   final String id;
   final String name;
   final String? avatar;
-  final int rank;
+  final double rank;
   final int top;
   final String provider;
